Update Line Persistent Menu css
