+++ ez.ai/client/src/components/ChatbotBuild/Preview/LinePreview/AdvancePreview/ListPreview.js
@@ -13,8 +13,8 @@ const ListPreview = ({
   return(
       <div className="main-preview">
         <div
-          className={now == i ? "main-content listbox-telegram now"
-                              : "main-content listbox-telegram"}
+          className={now == i ? "main-content listbox-line now"
+                              : "main-content listbox-line"}
           key={v.listContent + i}
           onClick={e => {
             e.stopPropagation();
@@ -24,21 +24,7 @@ const ListPreview = ({
             changeAvailableIcon("list");
           }}
         >
-          {" "}
-          <div>
-            <div className="listbox-header-telegram">Question</div>
-            <div className="listbox-question-telegram">
-              {v.listContent.question !== ""
-                ? v.listContent.question
-                : "(Ask a question)"}
-            </div>
-            {v.listContent.elem.map((i) => (
-              v.listContent.elem[i] && 
-                <div className="listbox-elem-telegram">
-                  {v.listContent.elem[i]}
-                </div>
-            ))}
-          </div>
+          <div className="listbox-header-line">고정메뉴</div>
         </div>
       <div
         className="tool-delete delete-listbox "
